new data creator module
